requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0